+++ skill.h
@@ -10,7 +10,7 @@ class Skill
 {
     private:
         string skillName;
-        int attackRate;
+        double attackRate;
         int biAttack;
     public:
         // コンストラクタ
@@ -22,9 +22,9 @@ class Skill
         // スキル名取得
         const string& getSkillName() const;
         // 攻撃力セット
-        void setAttackRate(int a);
+        void setAttackRate(double a);
         // 攻撃力取得
-        int getAttackRate() const;
+        double getAttackRate() const;
         // 攻撃回数セット
         void setBiAttack(int a);
         // 攻撃回数取得
